fix(rmm-eac2): fix MISRA C:2012 Rule 10.3 in pl011.c
Fixes
Event misra_c_2012_rule_10_3_violation:
Implicit conversion of "ch" from essential type
"character" to different or narrower essential
type "unsigned 8-bit int".

+++ drivers/pl011/src/pl011.c
@@ -64,7 +64,7 @@ int uart_init(uintptr_t base_addr,
 void uart_putc(char ch)
 {
 	uart_wait();
-	write8(ch, (void *)((RMM_UART_ADDR) + UARTDR));
+	write8((uint8_t)ch, (void *)((RMM_UART_ADDR) + UARTDR));
 }
 
 /* Serial output - called from printf */
